Introduce superclass
Three of our checks have very similar code. The simplest way of
extracting that common code is to use inheritance. We introduce an
UnexpectedRegexMatch class that holds the common code and use this as a
superclass for the UnexpectedReturnSymbol check.

Further commits will change other checks to inherit from
UnexpectedRegexMatch.