stop skipping common.greeting key
This key is used in our internal apps to test that locales are
configured and available, but it's not a standard rails key.

To confirm that, I created a new rails 5.0.3 app and the default en
locale has a single key: "hello".

I think it's best for this gem to allow those default keys to be tested,
and if the user wants to skip them they can add the key to
`exclude_keys`.