stop ignoring i18n keys in KeyUsageChecker
The check that confirms each key is used was hard coded to ignore any
keys that start with "i18n.".

This namespace is used by the i18n to lookup config for features like
pluralisation and transliteration. It's like that users will want these
keys ignored by the key usage check, but it's also possible there may be
other keys with the same prefix that they want to check.

I think it would be best for the check to assume all keys should be
checked, and the user can use `exclude_keys` or `exclude_scopes` to
explicitly choose the keys they want ignored.