Fix Greedy Regex
The regular expression matching the variables in the source was greedy
and producing spurious mismatching errors.

In particular if the source had the following:

This is the source [%{hello}](%{world}) it would require the translation to
have the variable :hello](world instead of two variables :hello and :world.

This commit fixes that.